Moved chart widget logic into separate abstract class
