+++ app/Filament/Widgets/RecentJitterChartWidget.php
@@ -2,50 +2,13 @@
 
 namespace App\Filament\Widgets;
 
-use App\Enums\ResultStatus;
-use App\Models\Result;
-use Filament\Widgets\ChartWidget;
-
-class RecentJitterChartWidget extends ChartWidget
+class RecentJitterChartWidget extends RecentChartWidget
 {
     protected static ?string $heading = 'Jitter';
 
-    protected int|string|array $columnSpan = 'full';
-
-    protected static ?string $maxHeight = '250px';
-
-    public ?string $filter = '24h';
-
-    protected function getPollingInterval(): ?string
-    {
-        return config('speedtest.dashboard_polling');
-    }
-
-    protected function getFilters(): ?array
-    {
-        return [
-            '24h' => 'Last 24h',
-            'week' => 'Last week',
-            'month' => 'Last month',
-        ];
-    }
-
     protected function getData(): array
     {
-        $results = Result::query()
-            ->select(['id', 'data', 'created_at'])
-            ->where('status', '=', ResultStatus::Completed)
-            ->when($this->filter == '24h', function ($query) {
-                $query->where('created_at', '>=', now()->subDay());
-            })
-            ->when($this->filter == 'week', function ($query) {
-                $query->where('created_at', '>=', now()->subWeek());
-            })
-            ->when($this->filter == 'month', function ($query) {
-                $query->where('created_at', '>=', now()->subMonth());
-            })
-            ->orderBy('created_at')
-            ->get();
+        $results = $this->getResults('data');
 
         return [
             'datasets' => [
@@ -85,32 +48,5 @@ class RecentJitterChartWidget extends ChartWidget
             ],
             'labels' => $results->map(fn ($item) => $item->created_at->timezone(config('app.display_timezone'))->format(config('app.chart_datetime_format'))),
         ];
-    }
-
-    protected function getOptions(): array
-    {
-        return [
-            'plugins' => [
-                'legend' => [
-                    'display' => true,
-                ],
-                'tooltip' => [
-                    'enabled' => true,
-                    'mode' => 'index',
-                    'intersect' => false,
-                    'position' => 'nearest',
-                ],
-            ],
-            'scales' => [
-                'y' => [
-                    'beginAtZero' => config('app.chart_begin_at_zero'),
-                ],
-            ],
-        ];
-    }
-
-    protected function getType(): string
-    {
-        return 'line';
     }
 }
