+++ app/Filament/Widgets/RecentPingChartWidget.php
@@ -2,51 +2,15 @@
 
 namespace App\Filament\Widgets;
 
-use App\Enums\ResultStatus;
 use App\Helpers\Average;
-use App\Models\Result;
-use Filament\Widgets\ChartWidget;
 
-class RecentPingChartWidget extends ChartWidget
+class RecentPingChartWidget extends RecentChartWidget
 {
     protected static ?string $heading = 'Ping (ms)';
 
-    protected int|string|array $columnSpan = 'full';
-
-    protected static ?string $maxHeight = '250px';
-
-    public ?string $filter = '24h';
-
-    protected function getPollingInterval(): ?string
-    {
-        return config('speedtest.dashboard_polling');
-    }
-
-    protected function getFilters(): ?array
-    {
-        return [
-            '24h' => 'Last 24h',
-            'week' => 'Last week',
-            'month' => 'Last month',
-        ];
-    }
-
     protected function getData(): array
     {
-        $results = Result::query()
-            ->select(['id', 'ping', 'created_at'])
-            ->where('status', '=', ResultStatus::Completed)
-            ->when($this->filter == '24h', function ($query) {
-                $query->where('created_at', '>=', now()->subDay());
-            })
-            ->when($this->filter == 'week', function ($query) {
-                $query->where('created_at', '>=', now()->subWeek());
-            })
-            ->when($this->filter == 'month', function ($query) {
-                $query->where('created_at', '>=', now()->subMonth());
-            })
-            ->orderBy('created_at')
-            ->get();
+        $results = $this->getResults('ping');
 
         return [
             'datasets' => [
@@ -74,33 +38,5 @@ class RecentPingChartWidget extends ChartWidget
             ],
             'labels' => $results->map(fn ($item) => $item->created_at->timezone(config('app.display_timezone'))->format(config('app.chart_datetime_format'))),
         ];
-    }
-
-    protected function getOptions(): array
-    {
-        return [
-            'plugins' => [
-                'legend' => [
-                    'display' => true,
-                ],
-                'tooltip' => [
-                    'enabled' => true,
-                    'mode' => 'index',
-                    'intersect' => false,
-                    'position' => 'nearest',
-                ],
-            ],
-            'scales' => [
-                'y' => [
-                    'beginAtZero' => config('app.chart_begin_at_zero'),
-                    'grace' => 2,
-                ],
-            ],
-        ];
-    }
-
-    protected function getType(): string
-    {
-        return 'line';
     }
 }
